test/api/requirements.txt
beautifulsoup4

flask

redis
wikipedia
requests


